backend/requirements.txt
nginx
fastapi
uvicorn
openai
python-dotenv
pdfplumber
aiofiles
python-multipart
stripe
